ref(web): restructure frontend directory and components (#4)
