+++ web/app/layout.tsx
@@ -1,25 +1,29 @@
-import type { Metadata } from 'next'
-import { GeistSans } from 'geist/font/sans'
-import { GeistMono } from 'geist/font/mono'
-import { Analytics } from '@vercel/analytics/next'
-import './globals.css'
+import type { Metadata } from "next";
+import { GeistSans } from "geist/font/sans";
+import { GeistMono } from "geist/font/mono";
+import { Analytics } from "@vercel/analytics/next";
+import "./globals.css";
 
 export const metadata: Metadata = {
-  title: 'TerraPeak - High-Performance Caching Proxy for Terraform Registry',
-  description: 'A high-performance caching proxy for Terraform Registry that accelerates provider downloads, supports intelligent storage backends, and offers proxy client/server modes for corporate environments.',
-}
+  title: "TerraPeak - High-Performance Caching Proxy for Terraform Registry",
+  description:
+    "A high-performance caching proxy for Terraform Registry that accelerates provider downloads, supports intelligent storage backends, and offers proxy client/server modes for corporate environments.",
+};
 
 export default function RootLayout({
   children,
 }: Readonly<{
-  children: React.ReactNode
+  children: React.ReactNode;
 }>) {
   return (
     <html lang="en">
-      <body className={`font-sans ${GeistSans.variable} ${GeistMono.variable}`} suppressHydrationWarning={true}>
+      <body
+        className={`font-sans ${GeistSans.variable} ${GeistMono.variable}`}
+        suppressHydrationWarning={true}
+      >
         {children}
         <Analytics />
       </body>
     </html>
-  )
+  );
 }
