+++ web/features/modules/components/ModuleList/ModuleFilterCard.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const ModuleFilterCard = () => {
+  return <div>ModuleFilterCard</div>;
+};
+
+export default ModuleFilterCard;
