+++ web/features/modules/components/ModuleList/ModuleFilterHeader.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const ModuleFilterHeader = () => {
+  return <div>ModuleFilterHeader</div>;
+};
+
+export default ModuleFilterHeader;
