+++ web/features/modules/components/ModuleList/ModuleItem.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const ModuleItem = () => {
+  return <div>ModuleItem</div>;
+};
+
+export default ModuleItem;
