+++ web/features/modules/data.ts
@@ -1,0 +1,125 @@
+export const MODULES_DATA = [
+  {
+    id: 1,
+    namespace: "terraform-aws-modules",
+    name: "vpc",
+    fullName: "terraform-aws-modules/vpc/aws",
+    description: "Terraform module which creates VPC resources on AWS",
+    tags: ["aws", "network", "vpc"],
+    stars: 2845,
+    version: "5.1.2",
+    provider: "hashicorp/aws",
+    versions: [
+      { version: "5.1.2", date: "2024-01-15" },
+      { version: "5.1.1", date: "2023-12-20" },
+      { version: "5.1.0", date: "2023-11-10" },
+    ],
+  },
+  {
+    id: 2,
+    namespace: "terraform-google-modules",
+    name: "kubernetes-engine",
+    fullName: "terraform-google-modules/kubernetes-engine/google",
+    description: "Modular and composable GKE cluster on Google Cloud Platform",
+    tags: ["gcp", "kubernetes", "gke"],
+    stars: 1523,
+    version: "28.0.0",
+    provider: "hashicorp/google",
+    versions: [
+      { version: "28.0.0", date: "2024-02-01" },
+      { version: "27.0.0", date: "2023-12-15" },
+    ],
+  },
+  {
+    id: 3,
+    namespace: "terraform-aws-modules",
+    name: "eks",
+    fullName: "terraform-aws-modules/eks/aws",
+    description:
+      "Terraform module to create an Elastic Kubernetes (EKS) cluster on AWS",
+    tags: ["aws", "kubernetes", "eks"],
+    stars: 3421,
+    version: "19.16.0",
+    provider: "hashicorp/aws",
+    versions: [
+      { version: "19.16.0", date: "2024-01-28" },
+      { version: "19.15.3", date: "2024-01-10" },
+    ],
+  },
+  {
+    id: 4,
+    namespace: "terraform-aws-modules",
+    name: "rds",
+    fullName: "terraform-aws-modules/rds/aws",
+    description: "Terraform module which creates RDS resources on AWS",
+    tags: ["aws", "database", "rds"],
+    stars: 1876,
+    version: "6.3.0",
+    provider: "hashicorp/aws",
+    versions: [
+      { version: "6.3.0", date: "2024-01-20" },
+      { version: "6.2.0", date: "2023-12-05" },
+    ],
+  },
+  {
+    id: 5,
+    namespace: "terraform-aws-modules",
+    name: "security-group",
+    fullName: "terraform-aws-modules/security-group/aws",
+    description:
+      "Terraform module which creates EC2 security group resources on AWS",
+    tags: ["aws", "security", "network"],
+    stars: 1234,
+    version: "5.1.0",
+    provider: "hashicorp/aws",
+    versions: [
+      { version: "5.1.0", date: "2024-01-12" },
+      { version: "5.0.0", date: "2023-11-20" },
+    ],
+  },
+  {
+    id: 6,
+    namespace: "terraform-google-modules",
+    name: "network",
+    fullName: "terraform-google-modules/network/google",
+    description: "Sets up a new VPC network on Google Cloud",
+    tags: ["gcp", "network", "vpc"],
+    stars: 987,
+    version: "8.0.0",
+    provider: "hashicorp/google",
+    versions: [
+      { version: "8.0.0", date: "2024-01-25" },
+      { version: "7.5.0", date: "2023-12-10" },
+    ],
+  },
+  {
+    id: 7,
+    namespace: "Azure",
+    name: "aks",
+    fullName: "Azure/aks/azurerm",
+    description: "Terraform module for deploying an AKS cluster on Azure",
+    tags: ["azure", "kubernetes", "aks"],
+    stars: 1654,
+    version: "7.5.0",
+    provider: "hashicorp/azurerm",
+    versions: [
+      { version: "7.5.0", date: "2024-02-05" },
+      { version: "7.4.0", date: "2024-01-15" },
+    ],
+  },
+  {
+    id: 8,
+    namespace: "terraform-aws-modules",
+    name: "s3-bucket",
+    fullName: "terraform-aws-modules/s3-bucket/aws",
+    description: "Terraform module which creates S3 bucket resources on AWS",
+    tags: ["aws", "storage", "s3"],
+    stars: 2156,
+    version: "3.15.1",
+    provider: "hashicorp/aws",
+    versions: [
+      { version: "3.15.1", date: "2024-01-30" },
+      { version: "3.15.0", date: "2024-01-05" },
+    ],
+  },
+];
